Update README.md
Because of the missing slash before the parameters, the site does not perceive the request correctly. The site adds the slash it needs to the request, but ignores the parameters passed. And when clicking on the link from the readme file, the user sees a blank page. This is not a problem_PROBLEM, but if it can be fixed, why not fix it.

The problem is reproduced on Chrome and Edge browsers

Signed-off-by: Maxim <therb1@mail.com>